refactor main function logic, move param logic to seperate file
